move(#132): batch 폴더 위치 변경 - batch 내에서 사용되는 코드 재사용 - batch는 독립적으로 동작하여 in/ 경로에 포함되어 있는것이 부적합하다고 판단.
