+++ semicolon/src/main/java/dukku/semicolon/boundedContext/settlement/batch/processor/ValidateSettlementProcessor.java
@@ -1,4 +1,4 @@
-package dukku.semicolon.boundedContext.settlement.in.batch.processor;
+package dukku.semicolon.boundedContext.settlement.batch.processor;
 
 import dukku.semicolon.boundedContext.settlement.entity.Settlement;
 import dukku.semicolon.shared.settlement.exception.SettlementValidationException;
