+++ semicolon/src/main/java/dukku/semicolon/boundedContext/settlement/batch/reader/FailedSettlementReaderConfig.java
@@ -1,8 +1,8 @@
-package dukku.semicolon.boundedContext.settlement.in.batch.reader;
+package dukku.semicolon.boundedContext.settlement.batch.reader;
 
 import dukku.semicolon.boundedContext.settlement.entity.Settlement;
 import dukku.semicolon.boundedContext.settlement.entity.type.SettlementStatus;
-import dukku.semicolon.boundedContext.settlement.in.batch.config.SettlementBatchProperties;
+import dukku.semicolon.boundedContext.settlement.batch.config.SettlementBatchProperties;
 import jakarta.persistence.EntityManagerFactory;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -17,40 +17,38 @@ import java.time.LocalDateTime;
 import java.util.Map;
 
 /**
- * Step 2: 금액 검증 대상 Settlement Reader
- * - PENDING 상태의 Settlement 조회
- * - 정산 예약일이 현재 시간 이전인 건만 조회
+ * 재처리 대상 Settlement Reader
+ * - FAILED 상태의 Settlement 조회
+ * - 1시간 전에 실패한 건들을 재처리하기 위함
  */
 @Slf4j
 @Configuration
 @RequiredArgsConstructor
-public class ValidateSettlementReader {
+public class FailedSettlementReaderConfig {
 
     private final EntityManagerFactory entityManagerFactory;
     private final SettlementBatchProperties batchProperties;
 
     @Bean(destroyMethod = "")
     @StepScope
-    public JpaPagingItemReader<Settlement> pendingSettlementForValidationReader(
+    public JpaPagingItemReader<Settlement> failedSettlementReader(
             @Value("#{jobParameters['now']}") LocalDateTime now
     ) {
         String jpql = """
                 SELECT s FROM Settlement s
                 WHERE s.settlementStatus = :status
-                AND s.settlementReservationDate <= :now
-                ORDER BY s.settlementReservationDate ASC
+                ORDER BY s.updatedAt ASC
                 """;
 
-        log.info("[Step 2] 금액 검증 대상 Settlement Reader 생성 - pageSize: {}, now: {}",
+        log.info("[Retry] FAILED Settlement Reader 생성 - pageSize: {}, now: {}",
                 batchProperties.getPageSize(), now);
 
         return new JpaPagingItemReaderBuilder<Settlement>()
-                .name("pendingSettlementForValidationReader")
+                .name("failedSettlementReader")
                 .entityManagerFactory(entityManagerFactory)
                 .queryString(jpql)
                 .parameterValues(Map.of(
-                        "status", SettlementStatus.PENDING,
-                        "now", now
+                        "status", SettlementStatus.FAILED
                 ))
                 .pageSize(batchProperties.getPageSize())
                 .saveState(true)
