+++ semicolon/src/main/java/dukku/semicolon/boundedContext/settlement/batch/scheduler/SettlementJobScheduler.java
@@ -1,6 +1,6 @@
-package dukku.semicolon.boundedContext.settlement.in.batch.scheduler;
+package dukku.semicolon.boundedContext.settlement.batch.scheduler;
 
-import dukku.semicolon.boundedContext.settlement.in.batch.config.SettlementBatchProperties;
+import dukku.semicolon.boundedContext.settlement.batch.config.SettlementBatchProperties;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.batch.core.job.Job;
