+++ semicolon/src/main/java/dukku/semicolon/boundedContext/settlement/batch/writer/DepositChargeWriter.java
@@ -1,4 +1,4 @@
-package dukku.semicolon.boundedContext.settlement.in.batch.writer;
+package dukku.semicolon.boundedContext.settlement.batch.writer;
 
 import dukku.semicolon.boundedContext.settlement.entity.Settlement;
 import dukku.semicolon.boundedContext.settlement.out.SettlementRepository;
@@ -9,27 +9,27 @@ import org.springframework.batch.infrastructure.item.ItemWriter;
 import org.springframework.stereotype.Component;
 
 /**
- * 재처리 Settlement 저장 Writer
- * - PENDING 상태로 복구된 Settlement 저장
+ * Step 2: 예치금 충전 완료된 Settlement 저장 Writer
+ * - SUCCESS 상태로 변경된 Settlement 저장
  */
 @Slf4j
 @Component
 @RequiredArgsConstructor
-public class RetrySettlementWriter implements ItemWriter<Settlement> {
+public class DepositChargeWriter implements ItemWriter<Settlement> {
 
     private final SettlementRepository settlementRepository;
 
     @Override
     public void write(Chunk<? extends Settlement> chunk) throws Exception {
-        log.info("[Retry Writer] PENDING 상태로 복구된 Settlement 저장 시작 - 건수: {}", chunk.size());
+        log.info("[Step 2 Writer] SUCCESS 상태 Settlement 저장 시작 - 건수: {}", chunk.size());
 
         for (Settlement settlement : chunk) {
             settlementRepository.save(settlement);
-            log.debug("[Retry Writer] Settlement 상태 업데이트 완료 - UUID: {}, 상태: {}",
+            log.debug("[Step 2 Writer] Settlement 상태 업데이트 완료 - UUID: {}, 상태: {}",
                     settlement.getUuid(),
                     settlement.getSettlementStatus());
         }
 
-        log.info("[Retry Writer] PENDING 상태로 복구된 Settlement 저장 완료 - 총 {}건 처리됨", chunk.size());
+        log.info("[Step 2 Writer] SUCCESS 상태 Settlement 저장 완료 - 총 {}건 처리됨", chunk.size());
     }
 }
